requirements.txt
numpy==2.4.4
openai==2.36.0
python-dotenv==1.2.2
pandas==3.0.3
datasets==4.8.5
ragas==0.4.3
